requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
click==8.1.8
Cython==3.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
docstring_parser==0.16
docutils==0.21.2
duckduckgo_search==7.3.2
edge-tts==6.1.19
elevenlabs==1.50.7
execnet==2.1.1
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpeg==1.4
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.24.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
imagesize==1.4.1
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-google-genai==2.0.9
langchain-google-vertexai==2.0.10
langchain-text-splitters==0.3.6
langsmith==0.3.7
Levenshtein==0.26.1
litellm==1.60.8
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mistune==3.1.1
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbsphinx==0.9.6
nest-asyncio==1.6.0
openai==1.61.1
orjson==3.10.15
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
primp==0.12.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pytest==8.3.4
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
referencing==0.36.2
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-autodoc-typehints==2.5.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.12.5
types-PyYAML==6.0.12.20241230
typing-inspect==0.9.0
tzdata==2025.1
uritemplate==4.1.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.44.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
zstandard==0.23.0
